fix pkg name
